Delete duplicates in code base. Optimized architecture for secure expansion of the project in the future.
